Fixes
Changes how rotation is calculated, adds null checks for if the skull isn't an actual player skull, and optimizes it so there are no useless packets or variables being set.
